+++ src/main/java/de/fraunhofer/isst/configmanager/communication/dataspaceconnector/DataspaceConnectorClient.java
@@ -8,16 +8,14 @@ import de.fraunhofer.iais.eis.Representation;
 import de.fraunhofer.iais.eis.Resource;
 import de.fraunhofer.iais.eis.ids.jsonld.Serializer;
 import de.fraunhofer.isst.configmanager.communication.clients.DefaultConnectorClient;
+import de.fraunhofer.isst.configmanager.communication.dataspaceconnector.model.BackendSource;
 import de.fraunhofer.isst.configmanager.communication.dataspaceconnector.model.ResourceRepresentation;
 import de.fraunhofer.isst.configmanager.util.OkHttpUtils;
 import lombok.AccessLevel;
 import lombok.experimental.FieldDefaults;
 import lombok.extern.slf4j.Slf4j;
-import okhttp3.Credentials;
-import okhttp3.HttpUrl;
-import okhttp3.OkHttpClient;
-import okhttp3.Request;
-import okhttp3.RequestBody;
+import okhttp3.*;
+import org.json.JSONObject;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnExpression;
 import org.springframework.stereotype.Service;
@@ -156,15 +154,15 @@ public class DataspaceConnectorClient implements DefaultConnectorClient {
     @Override
     public String getRequestedResourcesAsJsonString() throws IOException {
         final var baseConnectorNode = getJsonNodeOfBaseConnector();
-        final var offeredResourceNode = baseConnectorNode.findValue("ids:requestedResource");
-        return offeredResourceNode.toString();
+        final var requestedResourceNode = baseConnectorNode.findValue("ids:requestedResource");
+        return requestedResourceNode.toString();
     }
 
     private JsonNode getJsonNodeOfBaseConnector() throws IOException {
         final var builder = new Request.Builder();
         final var connectorUrl =
                 "https://" + dataSpaceConnectorHost + ":" + dataSpaceConnectorPort +
-                "/admin/api/connector";
+                        "/admin/api/connector";
         builder.header("Authorization", Credentials.basic(dataSpaceConnectorApiUsername,
                 dataSpaceConnectorApiPassword));
         builder.url(connectorUrl);
@@ -226,6 +224,65 @@ public class DataspaceConnectorClient implements DefaultConnectorClient {
         }
         final var body = Objects.requireNonNull(response.body()).string();
         return SERIALIZER.deserialize(body, BaseConnector.class);
+    }
+
+    @Override
+    public Resource getRequestedResource(String accessURL, String resourceId) throws IOException {
+        var builder = new Request.Builder();
+        var urlBuilder = new HttpUrl.Builder()
+                .scheme("https")
+                .host(dataSpaceConnectorHost)
+                .port(dataSpaceConnectorPort)
+                .addPathSegments("admin/api/request/description")
+                .addQueryParameter("recipient", accessURL)
+                .addQueryParameter("requestedResource", resourceId);
+        var url = urlBuilder.build();
+        log.info(url.toString());
+        builder.url(url);
+        builder.header("Authorization", Credentials.basic(dataSpaceConnectorApiUsername, dataSpaceConnectorApiPassword));
+        builder.post(RequestBody.create(null, new byte[0]));
+        var request = builder.build();
+        var response = client.newCall(request).execute();
+        if (!response.isSuccessful()) {
+            log.warn(String.format("Could not get BaseConnector from %s!", dataSpaceConnectorHost));
+        }
+        var body = response.body().string();
+        var splitBody = body.split("\n", 2);
+        String uuid = splitBody[0].substring(12);
+        log.info(uuid);
+        String resource = splitBody[1].substring(10);
+        log.info(resource);
+        return SERIALIZER.deserialize(resource, Resource.class);
+    }
+
+    @Override
+    public String requestContractAgreement(String recipientId, String requestedArtifactId, String contractOffer) throws IOException {
+        log.info("Request contract agreement with recipient: {} and artifact: {}", recipientId, requestedArtifactId);
+        var builder = new Request.Builder();
+        var urlBuilder = new HttpUrl.Builder()
+                .scheme("https")
+                .host(dataSpaceConnectorHost)
+                .port(dataSpaceConnectorPort)
+                .addPathSegments("admin/api/request/contract")
+                .addQueryParameter("recipient", recipientId)
+                .addQueryParameter("requestedArtifact", requestedArtifactId);
+        var url = urlBuilder.build();
+        log.info(url.toString());
+        builder.url(url);
+        builder.header("Authorization", Credentials.basic(dataSpaceConnectorApiUsername, dataSpaceConnectorApiPassword));
+        if (contractOffer != null && !contractOffer.isBlank()) {
+            builder.post(RequestBody.create(contractOffer, okhttp3.MediaType.parse("application/ld+json")));
+        } else {
+            builder.post(RequestBody.create(null, new byte[0]));
+        }
+        var request = builder.build();
+        var response = client.newCall(request).execute();
+        if (!response.isSuccessful()) {
+            log.warn("Could not request contract agreement");
+        }
+        var body = response.body().string();
+        log.info("Response: " + body);
+        return body;
     }
 
     @Override
@@ -509,26 +566,50 @@ public class DataspaceConnectorClient implements DefaultConnectorClient {
 
     @Override
     public String updateResource(final URI resourceID, final Resource resource) throws IOException {
-        log.info(String.format("---- updating resource at %s", dataSpaceConnectorHost));
+        log.info(String.format("---- [DataspaceConnectorClient updateResource] updating resource at %s", dataSpaceConnectorHost));
         final var mappedResource = dataSpaceConnectorResourceMapper.getMetadata(resource);
         final var path = resourceID.getPath();
         final var idStr = path.substring(path.lastIndexOf('/') + 1);
         final var resourceUUID = UUID.fromString(idStr);
-        final var resourceJsonLD = MAPPER.writeValueAsString(mappedResource);
+
+        BackendSource backendSource = new BackendSource();
+        try {
+            final var requestBackendBuilder = new Request.Builder();
+            requestBackendBuilder.url("https://" + dataSpaceConnectorHost + ":" + dataSpaceConnectorPort + "/admin" +
+                    "/api/resources/" + resourceUUID);
+            requestBackendBuilder.get();
+            requestBackendBuilder.header("Authorization", Credentials.basic(dataSpaceConnectorApiUsername,
+                    dataSpaceConnectorApiPassword));
+            final var request = requestBackendBuilder.build();
+            final var response = client.newCall(request).execute();
+
+            final var mapper = new ObjectMapper();
+            final var jsonTree = mapper.readTree(Objects.requireNonNull(response.body()).string());
+            JsonNode source = jsonTree.findValue("source");
+            backendSource.setType(BackendSource.Type.valueOf(source.get("type").asText().toUpperCase()));
+            backendSource.setUrl(URI.create(source.get("url").asText()));
+            backendSource.setUsername(source.get("username").asText());
+            backendSource.setPassword(source.get("password").asText());
+        } catch (Exception e) {
+            log.error(e.getMessage(), e);
+        }
+
+        mappedResource.getRepresentations().get(0).setSource(backendSource);
+        var resourceJsonLD = MAPPER.writeValueAsString(mappedResource);
+
         final var builder = new Request.Builder();
         builder.url("https://" + dataSpaceConnectorHost + ":" + dataSpaceConnectorPort + "/admin" +
                 "/api/resources/" + resourceUUID);
-        builder.put(RequestBody.create(resourceJsonLD, okhttp3.MediaType.parse("application/ld" +
-                "+json")));
+        builder.put(RequestBody.create(resourceJsonLD, okhttp3.MediaType.parse("application/ld+json")));
         builder.header("Authorization", Credentials.basic(dataSpaceConnectorApiUsername,
                 dataSpaceConnectorApiPassword));
         final var request = builder.build();
         final var response = client.newCall(request).execute();
         if (!response.isSuccessful()) {
-            log.warn(String.format("---- Updating Resource at %s failed!", dataSpaceConnectorHost));
+            log.warn(String.format("---- [DataspaceConnectorClient updateResource] Updating Resource at %s failed!", dataSpaceConnectorHost));
         }
         final var body = Objects.requireNonNull(response.body()).string();
-        log.info("---- Response: " + body);
+        log.info("---- [DataspaceConnectorClient updateResource] Response: " + body);
         return body;
     }