+++ src/main/java/de/fraunhofer/isst/configmanager/controller/ResourceContractUIController.java
@@ -54,15 +54,8 @@ public class ResourceContractUIController implements ResourceContractApi {
      * @return a suitable http response depending on success
      */
     @Override
-    //TODO use resourceService getResources
     public ResponseEntity<String> getResourceContract(final URI resourceId) {
         log.info(">> GET /resource/contract resourceId: " + resourceId);
-
-        if (configModelService.getConfigModel() == null ||
-                configModelService.getConfigModel().getConnectorDescription().getResourceCatalog() == null) {
-            return ResponseEntity.status(HttpStatus.NOT_FOUND).body("{\"error\":\"Could not find " +
-                    "any resources!\"}");
-        }
 
         final var contractOffer = resourceService.getResourceContract(resourceId);
         if (contractOffer != null) {